refactor: code locality
